use reducer hook added
